chore(*): split path & path_var into their own files
